feat(port): add getport and withImage
