+++ .gitignore
@@ -1,1 +1,2 @@
 /coverage.out
+/coverage.html
